Update dependencies and refactor import paths for consistency
