+++ pvm/ts/package.json
@@ -13,7 +13,7 @@
         "sign": "npx ts-node scripts/sign.ts"
     },
     "dependencies": {
-        "@bitcoinbrisbane/block52": "1.0.105",
+        "@bitcoinbrisbane/block52": "1.0.109",
         "@types/cors": "^2.8.17",
         "@types/node": "^22.10.3",
         "@types/ws": "^8.18.1",
